[analyzer] Fix sensitive argument logic in GenericTaintChecker
The semantics of taint sinks is that if ANY of the arguments is tainted, a
warning should be emmitted. Before this change, if there were multiple
arguments that are sensitive, and if the first arg is not tainted, but any of
the noninitial are tainted, a warning is not emitted. After this change the
correct semantics is reflected in code.

Differential Revision: https://reviews.llvm.org/D114706