[analyzer][clangsa] Add new option to alpha.security.cert.InvalidPtrChecker
The invalidation of pointer pointers returned by subsequent calls to genenv is
suggested by the POSIX standard, but is too strict from a practical point of
view. A new checker option 'InvalidatingGetEnv' is introduced, and is set to a
more lax default value, which does not consider consecutive getenv calls
invalidating.

Differential Revision: https://reviews.llvm.org/D154603